refactor(Handler): use passed argument only when needed
