+++ tomcat/src/main/java/nextstep/jwp/servlet/Handler.java
@@ -21,9 +21,16 @@ public class Handler {
 
     public Object handle(Object arg) {
         try {
-            return method.invoke(controller, arg);
+            return handleWithValidArgs(arg);
         } catch (InvocationTargetException | IllegalAccessException e) {
             throw new HttpException(e, HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
+
+    private Object handleWithValidArgs(Object arg) throws IllegalAccessException, InvocationTargetException {
+        if (method.getParameterCount() == 0) {
+            return method.invoke(controller);
+        }
+        return method.invoke(controller, arg);
     }
 }
